Fix rows response when error
